разбила на компоненты Dialogs + вынесла данные на уровень index
