+++ src/components/Dialogs/Dialogs.module.css
@@ -16,6 +16,7 @@
     margin: 10px;
     padding: 10px;
 }
-.messages .message{
 
+.item a.active {
+    color: brown;
 }
