+++ src/components/Dialogs/Message/Message.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import a from './../Dialogs.module.css'
+
+const Message = (props) => {
+    return <div className={a.message}>{props.message}</div>
+};
+export default Message
