+++ src/components/Profile/MyPosts/MyPosts.js
@@ -2,13 +2,8 @@ import React from "react";
 import a from "./MyPosts.module.css";
 import MyPost from "./MyPost/MyPost";
 
-let PostData = [
-    {id: 1, message: 'Хочу лето и тепло', like: 1},
-    {id: 2, message: 'Люблю осеннюю погоду', like: 2}
-];
-
-const MyPosts = () => {
-    let postElements = PostData.map((el) => {
+const MyPosts = (props) => {
+    let postElements = props.postData.map((el) => {
         return <MyPost message={el.message} id={el.id} like={el.likesCount}/>
     });
     return <div className={a.postsBlock}>
